backend/requirements.txt
fastapi
uvicorn
databases
pydantic-settings
aiosqlite
sqlalchemy
